Bind Wavs together

Bind Wavs together

tts_bind_wav(result)

Arguments

result
A data.frame from [text2speech::tts()].

Value

A `data.frame` with the same structure as that of tts

Note

As the data are split due to limits of the API, then this allows the text and the results to be harmonized