Text to Speech

Text to Speech

tts_google(
  text,
  output_format = c("mp3", "wav"),
  voice = "en-US-Standard-C",
  bind_audio = TRUE,
  ...
)

tts_amazon(
  text,
  output_format = c("mp3", "wav"),
  voice = "Joanna",
  bind_audio = TRUE,
  ...
)

tts_microsoft(
  text,
  output_format = c("mp3", "wav"),
  voice = "Microsoft Server Speech Text to Speech Voice (en-US, ZiraRUS)",
  bind_audio = TRUE,
  ...
)

tts_default_voice(service = c("amazon", "google", "microsoft"))

tts(
  text,
  output_format = c("mp3", "wav"),
  ...,
  service = c("amazon", "google", "microsoft"),
  bind_audio = TRUE
)

Arguments

text
A character vector of text to speak

output_format
Format of output files

voice
A full voice name that can be passed to the service, such as the argument voice for get_synthesis`` from \code{aws.polly}, or or [mscstts::ms_synthesize()] or the name` argument for googleLanguageR::gl_talk()

bind_audio
Should the text2speech::tts_bind_wav() be run on after the audio has been created, to ensure that the length of text and the number of rows is consistent? This affects the output format of some audio.

...
Additional arguments to text2speech::tts_google(), text2speech::tts_amazon(), or text2speech::tts_microsoft()

service
service to use

Value

A data.frame of text and wav files

Note

All functions have a voice`` argument fro a full voice name that can be passed to the service, such as voiceforget_synthesis`` from aws.polly

Examples

tts_default_voice("amazon")

#> [1] "Joanna"

tts_default_voice("google")

#> [1] "en-US-Standard-C"

tts_default_voice("microsoft")

#> [1] "Microsoft Server Speech Text to Speech Voice (en-US, ZiraRUS)"

if (requireNamespace("stringi", quietly = TRUE)) {
    set.seed(1)
    text = stringi::stri_rand_lipsum(10)
    text[3] = paste0(text[3:length(text)], collapse = " " )
    text = text[c(1,3)]
    nchar(text)
    if (tts_auth("google")) {
        res = tts(text, service = "google", bind_audio = FALSE)
        testthat::expect_equal(nrow(res), length(text) + 1)
        bound = tts_bind_wav(res)
        testthat::expect_equal(nrow(bound), length(text))
    }
}

#> 2020-01-31 13:07:54 -- Calling text-to-speech API: Lorem ipsum dolor sit amet, finibus torquent diam lobortis dolor ac eget neque sapien habitant tortor aliquam, nulla. Aliquet elit nec in quisque luctus amet in ac. Ante amet vestibulum, sit quis sit. Risus rhoncus sed urna dignissim sapien in eu aliquam nec! Eu mauris mauris, eu, facilisi, ipsum dui dapibus et ac in. Taciti, phasellus et augue lorem sapien, dictumst amet. Viverra ut erat non vel placerat consectetur, nunc. Augue ut ligula penatibus porttitor consequat tincidunt eros hac morbi magnis. Mattis pretium nunc finibus curae vestibulum erat mattis ut nibh.

#> 2020-01-31 13:07:55 -- Calling text-to-speech API: Pharetra et ultricies feugiat in dolor litora, in vestibulum. Neque maximus quis justo nibh primis ac, ut! Natoque purus sapien amet. In ultrices varius curae lobortis in nec. Risus sed, purus quis, netus nam, nec. Risus, commodo lorem montes iaculis erat ac vitae. Bibendum mi neque duis finibus, nec integer cum, tellus tempor sit. Sed fusce eu sed, netus vestibulum nec erat lorem mi. Metus cras in lectus commodo ultrices nullam diam. Penatibus eros integer facilisis libero fringilla quam. Sed porta nam sit malesuada pharetra donec ex venenatis aenean. Ligula eu nisl taciti curabitur facilisis mus habitasse lobortis, sed. Donec nulla ac in sapien donec. Amet porttitor erat montes ullamcorper, ipsum! Vel, ac hac ante felis scelerisque semper orci sed. Class rhoncus nisl et in euismod, venenatis ullamcorper eu turpis. Scelerisque class efficitur vitae malesuada nulla mus, nec. Magna eget et venenatis tortor, lacinia, eros placerat sodales ligula, dis. Malesuada sed blandit magna ut urna nascetur aenean lectus. Vitae mollis, ullamcorper convallis tristique dolor. Augue ante est viverra taciti non dictum, sociis sociis. Ligula velit dis et habitant. Turpis libero pharetra iaculis eros non turpis auctor in integer at, pharetra. Eu ipsum porttitor eget ex, sem, netus vivamus sapien tincidunt venenatis lacus tincidunt et. In commodo lacinia aptent pellentesque torquent vel. Posuere mauris sed praesent eleifend. Senectus, accumsan sagittis nibh velit odio et vel tellus dictum. Molestie sapien ultricies in porttitor egestas tellus himenaeos. Vel sapien lacinia pulvinar tortor venenatis tortor, tempor magna nibh, ac vestibulum nulla pulvinar mollis. Non arcu blandit tempor nisl justo sed litora. Ut dui euismod blandit justo suspendisse, congue nec penatibus diam. Taciti, ad dictum accumsan tristique ut justo urna. Primis, iaculis fermentum diam, ac pretium. Mi enim sed non justo, pulvinar sed lobortis aptent. Imperdiet quis integer. Donec, varius enim in porta, faucibus dictum in, iaculis vestibulum, nulla convallis. Ut in ligula tempus maximus nunc at ante. Diam proin ac accumsan neque ad faucibus eu purus sit mi mauris nascetur est condimentum. Nec, molestie at amet nec felis porta fusce aenean. Risus euismod sed hac, ex enim. Sed, quisque at platea class iaculis scelerisque diam bibendum nec. Luctus, vestibulum, lacus in amet. In primis turpis metus. Porttitor vitae arcu gravida ligula vivamus sed penatibus laoreet. Nunc accumsan ligula cum donec sed ligula. Nisl egestas mus nisl tempus sociosqu, inceptos

#> 2020-01-31 13:08:00 -- Calling text-to-speech API: vitae. Nam ligula, lectus laoreet euismod semper sed mauris. Nibh pellentesque porta elit, conubia sagittis. Mollis erat varius, ut fringilla montes. Tempor nostra eu vitae sem risus mi in. Malesuada nam ligula cum laoreet in facilisi, mauris sed quis consectetur eget. A sed quis. Duis sed, in erat sem elementum fermentum ut justo ridiculus. Viverra mauris sem, sed quis lacus, feugiat dictumst ipsum. Primis nascetur potenti augue, aliquet curabitur accumsan massa, ipsum, sed ut. Ut sed donec ac volutpat sed sed. Nec convallis ornare nunc sit, mauris in in non. Ut in tortor congue duis ut sed purus nec. Consectetur imperdiet hendrerit ipsum penatibus cubilia. In per euismod fermentum id tempor enim maximus sociosqu vitae nunc sodales. Dictum torquent lorem praesent, tincidunt at, litora in leo lacus magna. Lacinia facilisis conubia euismod et amet maecenas fringilla volutpat quis. Etiam sit sem sed egestas purus eu porttitor, pellentesque vitae egestas per, dictumst. Scelerisque habitasse pharetra mi, natoque non ut sed nisi efficitur. Ultricies eu nec, dictumst. Erat sit risus tempus ipsum platea. Urna eu mi luctus. Vel, eu condimentum ex nulla sit quis. Urna eleifend nisl, montes ac in. Tempor varius placerat ullamcorper primis dui ut hac. Erat fames, fusce vestibulum, eu. Nunc sed euismod amet sagittis nec ornare metus. Nunc turpis imperdiet volutpat diam donec. Parturient malesuada vel egestas arcu etiam non a nec nisi tempor sit eros. Massa tincidunt, dolor netus in libero purus class dignissim nec. Dictum eros suspendisse ut ac sociosqu dignissim nec urna taciti ipsum. Nec dis lorem mauris sed luctus eros ad tempor. Porta sed ante nisl donec nibh maecenas sed efficitur, viverra dui ac class. Leo diam iaculis vitae vestibulum imperdiet. Rhoncus, sed faucibus ullamcorper torquent commodo eget. Leo ac eu maecenas turpis justo netus nisl, a commodo sed eleifend. Congue dui pellentesque sodales pellentesque condimentum. Ac suscipit augue, proin facilisis ad nulla. Est nunc ipsum bibendum ac eros eros. Mi, amet nunc nullam cursus justo nec. Id ultricies nec, congue tortor aliquam nulla. Ultricies et vehicula ut, nam, bibendum adipiscing eget sed. Mauris ac nisl ut nulla odio, ipsum. Primis ante magnis nunc eleifend, eu. Mattis finibus purus lacus dapibus ut malesuada congue mauris in nisl turpis. Est nisl, inceptos amet non nisi sed! Ut mollis habitant at malesuada ipsum, montes. Cursus feugiat mauris in nibh feugiat habitant phasellus.